requirements.txt

Flask==2.3.2
requests==2.31.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
Flask-Limiter==3.5.0